% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDistance.R
\name{createDistance}
\alias{createDistance}
\title{Create distance matrix for a completely connected network}
\usage{
createDistance(coordMatrix)
}
\arguments{
\item{coordMatrix}{A matrix containing all the x and y coordinates of the network vertexes}
}
\value{
A square matrix containing the Euclidean distances between all vertexes, assuming that the network is completely connected.
}
\description{
Create distance matrix for a completely connected network
}
\examples{
coordMatrix <- matrix(c(0,10,15,20,30,30,15,15),ncol = 2)
createDistance(coordMatrix)

}
