% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidDistance.R
\name{euclidDistance}
\alias{euclidDistance}
\title{Euclidean distance between two points}
\usage{
euclidDistance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x-coordinate of point 1}

\item{y1}{y-coordinate of point 1}

\item{x2}{x-coordinate of point 2}

\item{y2}{y-coordinate of point 2}
}
\value{
The Euclidean distance between points 1 and 2 as a number
}
\description{
Euclidean distance between two points
}
