% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plotting a network consisting of edges and vertexes}
\usage{
plotNetwork(edgeMatrix, coordMatrix)
}
\arguments{
\item{edgeMatrix}{A matrix containing zeros and ones if an edge between two vertexes is absent or not}

\item{coordMatrix}{A data frame containing the x and y coordinates of each vertex of the network}
}
\value{
A plot of the connected network
edgeMatrix <- matrix(0, nrow = 4, ncol = 4)
edgeMatrix[,1] <- c(0,1,0,0)
edgeMatrix[,2] <- c(1,0,1,1)
edgeMatrix[,3] <- c(0,1,0,0)
edgeMatrix[,4] <- c(0,1,0,0)
coordMatrix <- matrix(c(0,10,15,20,30,30,15,15),ncol = 2)
plotNetwork(edgeMatrix, coordMatrix)
}
\description{
Plotting a network consisting of edges and vertexes
}
