% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primDistance.R
\name{primDistance}
\alias{primDistance}
\title{Minimum spanning tree using Prim's algorithm}
\usage{
primDistance(distMatrix)
}
\arguments{
\item{distMatrix}{A square matrix containing the distances between all vertexes of a network}
}
\value{
A matrix with rows describing which vertex is connected to which other vertex.
}
\description{
Minimum spanning tree using Prim's algorithm
}
\examples{
distMatrix <- matrix(c(0,10,20,30,10,0,40,60,20,40,0,30,30,60,30,0),
nrow = 4, ncol = 4, byrow = TRUE)
primDistance(distMatrix)

}
