% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCoordinates.R
\name{randomCoordinates}
\alias{randomCoordinates}
\title{Create random coordinates for network vertexes}
\usage{
randomCoordinates(nNodes, xMax, xMin, yMax, yMin)
}
\arguments{
\item{nNodes}{The number of vertexes/ nodes in the network}

\item{xMax}{The maximum x-coordinate of the nodes in the network}

\item{xMin}{The minimum x-coordinate of the nodes in the network}

\item{yMax}{The maximum y-coordinate of the nodes in the network}

\item{yMin}{The minimum y-coordinate of the nodes in the network}
}
\value{
A data frame with dimensions nNodes x 2 containing the x and y coordinates of the network's vertexes
}
\description{
Create random coordinates for network vertexes
}
\examples{
nNodes <- 10
xMax <- 2000
xMin <- 0
yMax <- 3000
yMin <- 200
randomCoordinates(nNodes, xMax, xMin, yMax, yMin)

}
